+++ src/rgw/rgw_loadgen.h
@@ -69,7 +69,7 @@ public:
     return env;
   }
 
-  int complete_request() override;
+  size_t complete_request() override;
 };
 
 #endif
